# openapi.yaml
openapi: 3.0.3
info:
  title: FinMate API
  description: |
    API Contract for FinMate, a Personal Finance & Lifestyle Companion.
    Provides endpoints for Auth, Users, Groups, Expenses, Settlements, Notes, Goals, and Import/Export.

    ### Versioning Strategy
    URL-based versioning is used. All paths start with `/api/v1`.

    ### Pagination & Filtering
    Paginated GET endpoints support standard query parameters:
    * `page`: 1-based page index (default: `1`)
    * `limit`: Page size (default: `20`, max: `100`)
    * `cursor`: Cursor string (for cursor-based pagination on rapidly updated lists like Expenses and Audit Logs)
  version: 1.0.0
servers:
  - url: /api/v1
    description: Local Dev Server
paths:
...
  /goals:
    get:
      summary: List the caller's goals (owner-scoped, priority-ordered)
      tags: [Goals]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Goals retrieved (title returned as ciphertext)
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      summary: Create a goal (born-E2EE title; requires recovery — REC-1)
      tags: [Goals]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, encryptedContentKey, targetAmount, currency]
              properties:
                title:
                  type: string
                  description: E2EE ciphertext of the title — the server never decrypts it.
                encryptedContentKey:
                  type: string
                  description: Owner's RSA-wrapped per-goal content key (born-E2EE).
                targetAmount: { type: number, format: float }
                savedAmount: { type: number, format: float }
                currency: { type: string, maxLength: 3 }
                targetDate: { type: string, format: date }
                priority: { type: integer, minimum: 0 }
      responses:
        '201':
          description: Goal created
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '409':
          description: Recovery not configured (errorCode REC_RECOVERY_REQUIRED)
  /goals/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: Get a goal (owner-scoped; title returned as ciphertext)
      tags: [Goals]
      security:
        - bearerAuth: []
      responses:
        '200': { description: Goal retrieved }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
    patch:
      summary: Update a goal (owner-scoped; optimistic version)
      tags: [Goals]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [version]
              properties:
                version: { type: integer }
                title: { type: string }
                encryptedContentKey: { type: string }
                targetAmount: { type: number }
                savedAmount: { type: number }
                currency: { type: string, maxLength: 3 }
                targetDate: { type: string, format: date }
                priority: { type: integer, minimum: 0 }
                status:
                  { type: string, enum: [active, achieved, paused, cancelled] }
      responses:
        '200': { description: Goal updated }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
        '412': { $ref: '#/components/responses/PreconditionFailedError' }
    delete:
      summary: Delete a goal (owner-scoped; crypto-shreds the wrapped key)
      tags: [Goals]
      security:
        - bearerAuth: []
      responses:
        '200': { description: Goal deleted }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
  /goals/{id}/projection:
    get:
      summary: Deterministic goal projection (Goal Engine V1, numeric only)
      tags: [Goals]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: assumedMonthlyContribution
          in: query
          required: false
          schema: { type: number }
      responses:
        '200':
          description: Projection (status ok/insufficient_data/invalid_goal; no free-text)
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    ValidationError:
      description: Input validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 400
            timestamp: '2026-06-09T17:15:00.000Z'
            path: '/api/v1/auth/register'
            errorCode: 'VAL_INVALID_INPUT'
            message: 'Input validation failed'
            details:
              - field: 'email'
                issue: 'must be a valid email address'
            retryable: false
    UnauthorizedError:
      description: Authentication failed or token missing/expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 401
            timestamp: '2026-06-09T17:15:00.000Z'
            path: '/api/v1/users/me'
            errorCode: 'AUTH_TOKEN_EXPIRED'
            message: 'JWT access token has expired'
            retryable: false
...
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 404
            timestamp: '2026-06-09T17:15:00.000Z'
            path: '/api/v1/expenses/f81d4fae-7dec-11d0-a765-00a0c91e6bf6'
            errorCode: 'RES_NOT_FOUND'
            message: 'Expense with ID f81d4fae-7dec-11d0-a765-00a0c91e6bf6 does not exist.'
            retryable: false
...
    PreconditionFailedError:
      description: State conflict or optimistic lock version mismatch
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 412
            timestamp: '2026-06-09T17:15:00.000Z'
            path: '/api/v1/expenses/f81d4fae-7dec-11d0-a765-00a0c91e6bf6'
            errorCode: 'CON_VERSION_CONFLICT'
            message: 'The resource version has changed. Please fetch the latest state and retry.'
            retryable: true
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - statusCode
        - timestamp
        - path
        - errorCode
        - message
        - retryable
      properties:
        statusCode:
          type: integer
        timestamp:
          type: string
          format: date-time
        path:
          type: string
        errorCode:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: object
            required:
              - field
              - issue
            properties:
              field:
                type: string
              issue:
                type: string
        retryable:
          type: boolean
...